requirements.txt
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
dnspython==2.8.0
docopt==0.6.2
Flask==2.3.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
greenlet==3.2.4
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
psycopg2-binary==2.9.7
pymongo==4.15.3
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==2.3.6
